NEWSUBCREATE add name copy enhancement & other updates
Change overview
1.	Allow users to select existing names to be copied to the new share record being created.
a.	Share and loan name types eligible to be copied over are controlled by the new configuration program settings
i.	Copy share names flag & eligible name types
ii.	Copy loan names flag & loan name types
b.	While the program builds an internal list of name records which can be copied, based upon the config settings, it is also checking for duplicates (based upon the SSN field in the name records.)
i.	If the program encounters multiple name records with the same SSN, the first one found will be presented to the member as an eligible option to copy.
2.	Add “testing” function to the program
a.	Allow the CU to turn on test mode and set up to 10 test accounts.
b.	With test mode on if a member is one of the test accounts, then they will be able to test some of the new functionality whereas other members will be limited to current functionality.
c.	New functionality the test accounts will have access to are:
i.	Ability to copy existing name records to the new share being created
ii.	Ability to select a new share from up to 20 defined groups instead of 10.
3.	Add CU defined error code descriptions
a.	These changes greatly reduce the possibility of the member getting an “oops” error. Instead, for most error conditions which can occur before the new share is created, the member would be presented with an error message better clarifying what the error was. These 11 error messages can be defined by the CU via the configuration program.
4.	Added ability for the CU to add comment lines to the Terms or Conditions Letter files.
5.	Corrected a few JSON issues
6.	Corrected issues with the configuration program’s form layout keeping it from properly and accurately being rendered.
7.	Added ability to force error conditions (1 through 11) to facilitate testing program error conditions
a.	Requires configuration program to be in debug mode and error code entered into the test error code input field
8.	Corrected how name creation error and share transfer errors were being handled. (they weren’t)