Update BANNO.M2MTRANSFERS.RPT.BATCH
Added check for transfers/transactions to cross accounts member has access to and removed them from the report. This required including the BUILDSHARELOANLIST procedures instead of including them due to the procedures referring to the @SYMXINSTANCEID and the @CLIENTNUMNUMBER variables not available in batch mode.

Added some debug info to the report when the debug flag is on

Corrected issue with Day 1, Day 2 calculation / report output.